+++ front-packages/tailored-export/src/feature/components/SourceDetails/Text/__mocks__/CleanHTMLTags.tsx
@@ -1,0 +1,21 @@
+import React from 'react';
+import {CleanHTMLTagsOperation} from '../CleanHTMLTags';
+
+const CleanHTMLTags = ({
+  onOperationChange,
+}: {
+  onOperationChange: (updatedOperation: CleanHTMLTagsOperation) => void;
+}) => (
+  <button
+    onClick={() =>
+      onOperationChange({
+        type: 'clean_html_tags',
+        value: true,
+      })
+    }
+  >
+    Clean HTML tags
+  </button>
+);
+
+export {CleanHTMLTags};