+++ src/Akeneo/Pim/Permission/Bundle/DependencyInjection/AkeneoPimPermissionExtension.php
@@ -27,6 +27,7 @@ class AkeneoPimPermissionExtension extends Extension
         $loader->load('array_converters.yml');
         $loader->load('checkers.yml');
         $loader->load('cli_commands.yml');
+        $loader->load('completeness.yml');
         $loader->load('controllers.yml');
         $loader->load('entities.yml');
         $loader->load('event_subscribers.yml');