+++ src/Akeneo/Platform/Bundle/TailoredExportBundle/src/Infrastructure/Hydrator/Value/AssociationTypeValueHydrator.php
@@ -17,7 +17,6 @@ use Akeneo\Pim\Enrichment\Component\Product\Model\AssociationInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\GroupInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelInterface;
-use Akeneo\Platform\TailoredExport\Application\Common\Source\AssociationTypeSource;
 use Akeneo\Platform\TailoredExport\Application\Common\SourceValue\QuantifiedAssociation;
 use Akeneo\Platform\TailoredExport\Application\Common\SourceValue\QuantifiedAssociationsValue;
 use Akeneo\Platform\TailoredExport\Application\Common\SourceValue\SimpleAssociationsValue;
@@ -25,10 +24,15 @@ use Akeneo\Platform\TailoredExport\Application\Common\SourceValue\SourceValueInt
 
 final class AssociationTypeValueHydrator
 {
-    public function hydrate(ProductInterface $product, string $associationTypeCode, bool $isQuantified): SourceValueInterface
-    {
+    public function hydrate(
+        $productOrProductModel,
+        string $associationTypeCode,
+        bool $isQuantified
+    ): SourceValueInterface {
+        $this->checkProductOrProductModelEntity($productOrProductModel);
+
         if ($isQuantified) {
-            $normalizedQuantifiedAssociations = $product->getQuantifiedAssociations()->normalize()[$associationTypeCode] ?? [];
+            $normalizedQuantifiedAssociations = $productOrProductModel->getQuantifiedAssociations()->normalize()[$associationTypeCode] ?? [];
 
             return new QuantifiedAssociationsValue(
                 $this->getProductQuantifiedAssociations($normalizedQuantifiedAssociations),
@@ -37,15 +41,17 @@ final class AssociationTypeValueHydrator
         }
 
         return new SimpleAssociationsValue(
-            $this->getAssociatedProductIdentifiers($product, $associationTypeCode),
-            $this->getAssociatedProductModelCodes($product, $associationTypeCode),
-            $this->getAssociatedGroupCodes($product, $associationTypeCode),
+            $this->getAssociatedProductIdentifiers($productOrProductModel, $associationTypeCode),
+            $this->getAssociatedProductModelCodes($productOrProductModel, $associationTypeCode),
+            $this->getAssociatedGroupCodes($productOrProductModel, $associationTypeCode),
         );
     }
 
-    private function getAssociatedProductIdentifiers(ProductInterface $product, string $associationTypeCode): array
+    private function getAssociatedProductIdentifiers($productOrProductModel, string $associationTypeCode): array
     {
-        $association = $this->getAssociationForTypeCode($product, $associationTypeCode);
+        $this->checkProductOrProductModelEntity($productOrProductModel);
+
+        $association = $this->getAssociationForTypeCode($productOrProductModel, $associationTypeCode);
         $associatedProducts = $association ? $association->getProducts()->toArray() : [];
 
         return array_map(
@@ -54,9 +60,11 @@ final class AssociationTypeValueHydrator
         );
     }
 
-    private function getAssociatedProductModelCodes(ProductInterface $product, string $associationTypeCode): array
+    private function getAssociatedProductModelCodes($productOrProductModel, string $associationTypeCode): array
     {
-        $association = $this->getAssociationForTypeCode($product, $associationTypeCode);
+        $this->checkProductOrProductModelEntity($productOrProductModel);
+
+        $association = $this->getAssociationForTypeCode($productOrProductModel, $associationTypeCode);
         $associatedProductModels = $association ? $association->getProductModels()->toArray() : [];
 
         return array_map(
@@ -65,9 +73,11 @@ final class AssociationTypeValueHydrator
         );
     }
 
-    private function getAssociatedGroupCodes(ProductInterface $product, string $associationTypeCode): array
+    private function getAssociatedGroupCodes($productOrProductModel, string $associationTypeCode): array
     {
-        $association = $this->getAssociationForTypeCode($product, $associationTypeCode);
+        $this->checkProductOrProductModelEntity($productOrProductModel);
+
+        $association = $this->getAssociationForTypeCode($productOrProductModel, $associationTypeCode);
         $associatedGroups = $association ? $association->getGroups()->toArray() : [];
 
         return array_map(
@@ -76,9 +86,11 @@ final class AssociationTypeValueHydrator
         );
     }
 
-    private function getAssociationForTypeCode(ProductInterface $product, string $typeCode): ?AssociationInterface
+    private function getAssociationForTypeCode($productOrProductModel, string $typeCode): ?AssociationInterface
     {
-        foreach ($product->getAllAssociations() as $association) {
+        $this->checkProductOrProductModelEntity($productOrProductModel);
+
+        foreach ($productOrProductModel->getAllAssociations() as $association) {
             if ($association->getAssociationType()->getCode() === $typeCode) {
                 return $association;
             }
@@ -127,5 +139,15 @@ final class AssociationTypeValueHydrator
             ),
             $normalizedProductModelQuantifiedAssociations
         );
+    }
+
+    private function checkProductOrProductModelEntity($productOrProductModel): void
+    {
+        if (
+            !$productOrProductModel instanceof ProductInterface
+            && !$productOrProductModel instanceof ProductModelInterface
+        ) {
+            throw new \InvalidArgumentException('Cannot hydrate this entity');
+        }
     }
 }