+++ src/Akeneo/Platform/Bundle/TailoredExportBundle/src/Infrastructure/Validation/Source/Scalar/ScalarSourceValidator.php
@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace Akeneo\Platform\TailoredExport\Infrastructure\Validation\Source\Scalar;
 
+use Akeneo\Platform\TailoredExport\Infrastructure\Validation\Operation\CleanHTMLTagsOperationConstraint;
 use Akeneo\Platform\TailoredExport\Infrastructure\Validation\Operation\DefaultValueOperationConstraint;
 use Akeneo\Platform\TailoredExport\Infrastructure\Validation\Source\SourceConstraintProvider;
 use Symfony\Component\Validator\Constraint;
@@ -30,6 +31,7 @@ class ScalarSourceValidator extends ConstraintValidator
         $sourceConstraintFields['selection'] = new Collection(['fields' => ['type' => new EqualTo(['value' => 'code'])]]);
         $sourceConstraintFields['operations'] = new Collection(['fields' => [
             'default_value' => new Optional(new DefaultValueOperationConstraint()),
+            'clean_html_tags' => new Optional(new CleanHTMLTagsOperationConstraint()),
         ]]);
 
         $violations = $validator->validate($source, new Collection(['fields' => $sourceConstraintFields]));