+++ upgrades/test_schema/Version_5_0_20200706112252_unauthorize_pimee_catalog_rule_rule_create_permissions_integration.php
@@ -126,7 +126,7 @@ class Version_5_0_20200706112252_unauthorize_pimee_catalog_rule_rule_create_perm
     {
         $migration = (new \ReflectionClass($this))->getShortName();
         $migration = str_replace('_Integration', '', $migration);
-        $migration = str_replace('Version', '', $migration);
+        $migration = str_replace('Version', 'Pim\Upgrade\Schema\Version', $migration);
 
         return $migration;
     }